requirements.txt
-r weather_project/requirements.txt
-r affirmations_project/requirements.txt
python-dotenv
